Downgraded from JUnit 5=>4, Upgraded Kotlin 1.3.61 => 1.3.70, coroutines-core 1.3.2 => 1.3.3, added unit test to measure the performance of creating the invoice [380-436]ms + generating QR code [35-50]ms When doing a warm up or 2nd call invoice=180ms qrCode=17ms
