+++ app/build.gradle
@@ -1,7 +1,6 @@
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
-apply plugin: "de.mannodermaus.android-junit5"
 
 tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
     kotlinOptions.jvmTarget = "1.8"
@@ -53,7 +52,7 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.3.2"
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.3.3"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.2"
     implementation "androidx.lifecycle:lifecycle-runtime-ktx:2.2.0"
     implementation 'androidx.core:core-ktx:1.2.0'
@@ -81,24 +80,11 @@ dependencies {
     implementation 'com.github.pokkst:bitcoincashj:0.17.5.2'
     implementation "com.amplitude:android-sdk:2.24.2"
 
-    testImplementation "org.junit.jupiter:junit-jupiter-api:5.5.2"
-    testImplementation "org.junit.jupiter:junit-jupiter-params:5.5.2"
-    testRuntimeOnly "org.junit.jupiter:junit-jupiter-engine:5.5.2"
+    testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test.ext:junit:1.1.1'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
     androidTestImplementation 'androidx.test.espresso:espresso-intents:3.2.0'
-//    androidTestImplementation "com.android.support.test:runner:1.0.2"
-//    androidTestImplementation "com.android.support.test.espresso:espresso-core:3.0.2"
-//    androidTestImplementation "com.android.support.test.espresso:espresso-intents:3.0.2"
-    androidTestRuntimeOnly "de.mannodermaus.junit5:android-instrumentation-test-runner:0.2.2"
-    // Add the Android Instrumentation Test dependencies to the product flavor only
-    // androidTestExperimentalImplementation "org.junit.jupiter:junit-jupiter-api:$junit_jupiter_version"
-    // androidTestExperimentalImplementation "de.mannodermaus.junit5:android-instrumentation-test:0.2.2"
-    // Runtime dependencies for Android Instrumentation Tests
-    // androidTestExperimentalRuntimeOnly "org.junit.jupiter:junit-jupiter-engine:1.3.1"
-    // androidTestExperimentalRuntimeOnly "org.junit.platform:junit-platform-runner:1.3.1"
-
-    implementation('com.crashlytics.sdk.android:crashlytics:2.10.1@aar') {
-        transitive = true;
-    }
+    androidTestImplementation "com.android.support.test:runner:1.0.2"
+    androidTestImplementation "com.android.support.test.espresso:espresso-core:3.0.2"
+    androidTestImplementation "com.android.support.test.espresso:espresso-intents:3.0.2"
 }
