+++ app/src/test/java/org/bitcoindotcom/bchprocessor/bip70/model/InvoiceStatusTestJson.kt
@@ -1,9 +1,9 @@
 package org.bitcoindotcom.bchprocessor.bip70.model
 
 import org.bitcoindotcom.bchprocessor.bip70.GsonHelper
-import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Assertions.assertTrue
-import org.junit.jupiter.api.Test
+import org.junit.Assert.assertEquals
+import org.junit.Assert.assertTrue
+import org.junit.Test
 
 internal class InvoiceStatusTestJson {
     @Test
