+++ build.gradle
@@ -6,11 +6,10 @@ buildscript {
         jcenter()
         maven { url "https://jitpack.io" }
     }
-    ext.kotlin_version = '1.3.61'
+    ext.kotlin_version = '1.3.70'
     dependencies {
         classpath 'com.android.tools.build:gradle:3.6.1'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-        classpath "de.mannodermaus.gradle.plugins:android-junit5:1.5.2.0"
     }
 }
